requirements.txt
seaborn
pubchempy
pandas
rdkit
matplotlib
scikit-learn
pdfplumber 
googlesearch-python
